backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
requests
transformers
bibtexparser
PyMuPDF
torch
python-multipart
openpyxl
pymongo
pyjwt
cryptography
xlsxwriter